Update on migration files
Inclusion of check if exists in the Up method

+++ database/migrations/2017_02_08_073456_create_offices_table.php
@@ -12,27 +12,29 @@ class CreateOfficesTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::create('offices', function(Blueprint $table)
-		{
-			$table->increments('id');
-			$table->string('code',20)->unique();
-			$table->string('name');	
-			$table->string('description')->nullable()->default(null);
-			$table->char('abbreviation', 6)->nullable()->default(null);
-			$table->string('head')->nullable();
-			$table->string('head_title')->nullable()->default(null);
-			$table->string('status')
-				->nullable()
-				->default('In Service');
-			$table->integer('head_office')->unsigned()->nullable()->default(null);
-			$table->foreign('head_office')
-					->references('id')
-					->on('offices')
-					->onUpdate('cascade')
-					->onDelete('cascade');
-			$table->timestamps();
-			$table->softDeletes();		
-		});
+		if(!Schema::hasTable('offices')){
+			Schema::create('offices', function(Blueprint $table)
+			{
+				$table->increments('id');
+				$table->string('code',20)->unique();
+				$table->string('name');	
+				$table->string('description')->nullable()->default(null);
+				$table->char('abbreviation', 6)->nullable()->default(null);
+				$table->string('head')->nullable();
+				$table->string('head_title')->nullable()->default(null);
+				$table->string('status')
+					->nullable()
+					->default('In Service');
+				$table->integer('head_office')->unsigned()->nullable()->default(null);
+				$table->foreign('head_office')
+						->references('id')
+						->on('offices')
+						->onUpdate('cascade')
+						->onDelete('cascade');
+				$table->timestamps();
+				$table->softDeletes();		
+			});
+		} 
 	}
 
 	/**
