+++ database/migrations/2017_07_21_073957_create_supplies_table.php
@@ -12,27 +12,29 @@ class CreateSuppliesTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::create('supplies', function(Blueprint $table)
-		{
-			$table->increments('id');
-			$table->string('stocknumber')->unique();
-			$table->string('entity_name',200)->default('Polytechnic University of the Philippines');
-			$table->integer('category_id')->unsigned()->nullable();
-			$table->foreign('category_id')
-					->references('id')
-					->on('categories')
-					->onUpdate('cascade')
-					->onDelete('cascade');
-            $table->string('details');		
-            $table->integer('unit_id')->unsigned()->nullable();
-            $table->foreign('unit_id')
-            		->references('id')
-            		->on('units')
-            		->onUpdate('cascade')
-            		->onDelete('cascade');
-            $table->integer('reorderpoint')->nullable();
-			$table->timestamps();
-		});
+		if(!Schema::hasTable('supplies')){
+			Schema::create('supplies', function(Blueprint $table)
+			{
+				$table->increments('id');
+				$table->string('stocknumber')->unique();
+				$table->string('entity_name',200)->default('Polytechnic University of the Philippines');
+				$table->integer('category_id')->unsigned()->nullable();
+				$table->foreign('category_id')
+						->references('id')
+						->on('categories')
+						->onUpdate('cascade')
+						->onDelete('cascade');
+				$table->string('details');		
+				$table->integer('unit_id')->unsigned()->nullable();
+				$table->foreign('unit_id')
+						->references('id')
+						->on('units')
+						->onUpdate('cascade')
+						->onDelete('cascade');
+				$table->integer('reorderpoint')->nullable();
+				$table->timestamps();
+			});
+		} 
 	}
 
 	/**
