+++ database/migrations/2017_08_04_141643_create_purchase_orders_table.php
@@ -13,21 +13,23 @@ class CreatePurchaseOrdersTable extends Migration
      */
     public function up()
     {
-        Schema::create('purchaseorders', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('number',100)->unique();
-            $table->date('date_received');
-            $table->string('details')->nullable();
-            $table->integer('created_by')->nullable();
-            $table->integer('supplier_id')->unsigned()->nullable();
-            $table->foreign('supplier_id')
-                    ->references('id')
-                    ->on('suppliers')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
-            $table->string('status')->nullable();
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('purchaseorders')){
+            Schema::create('purchaseorders', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('number',100)->unique();
+                $table->date('date_received');
+                $table->string('details')->nullable();
+                $table->integer('created_by')->nullable();
+                $table->integer('supplier_id')->unsigned()->nullable();
+                $table->foreign('supplier_id')
+                        ->references('id')
+                        ->on('suppliers')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');
+                $table->string('status')->nullable();
+                $table->timestamps();
+            });
+        } 
     }
 
     /**
