+++ database/migrations/2017_10_22_062016_create_requests_supplies_table.php
@@ -13,26 +13,28 @@ class CreateRequestsSuppliesTable extends Migration
      */
     public function up()
     {
-        Schema::create('requests_supplies', function (Blueprint $table) {
-            $table->integer('supply_id')->unsigned();
-            $table->foreign('supply_id')
-                    ->references('id')
-                    ->on('supplies')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
-            $table->integer('request_id')->unsigned();
-            $table->foreign('request_id')
-                    ->references('id')
-                    ->on('requests')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
-            $table->integer('quantity_requested')->default(0);
-            $table->integer('quantity_issued')->nullable();
-            $table->integer('quantity_released')->nullable();
-            $table->string('comments')->nullable();
-            $table->primary(['request_id', 'supply_id'], 'requests_supplies_primary');
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('requests_supplies')){
+            Schema::create('requests_supplies', function (Blueprint $table) {
+                $table->integer('supply_id')->unsigned();
+                $table->foreign('supply_id')
+                        ->references('id')
+                        ->on('supplies')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');
+                $table->integer('request_id')->unsigned();
+                $table->foreign('request_id')
+                        ->references('id')
+                        ->on('requests')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');
+                $table->integer('quantity_requested')->default(0);
+                $table->integer('quantity_issued')->nullable();
+                $table->integer('quantity_released')->nullable();
+                $table->string('comments')->nullable();
+                $table->primary(['request_id', 'supply_id'], 'requests_supplies_primary');
+                $table->timestamps();
+            });
+        } 
     }
 
     /**
