+++ database/migrations/2017_11_08_144437_create_requests_comments_table.php
@@ -13,24 +13,26 @@ class CreateRequestsCommentsTable extends Migration
      */
     public function up()
     {
-        Schema::create('requests_comments', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('request_id')->unsigned();
-            $table->foreign('request_id')
-                    ->references('id')
-                    ->on('requests')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
-            $table->integer('user_id')->unsigned();
-            $table->foreign('user_id')
-                    ->references('id')
-                    ->on('users')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
-            $table->timestamps();
-            $table->longtext('details');
-            $table->softDeletes();
-        });
+        if(!Schema::hasTable('requests_comments')){
+            Schema::create('requests_comments', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('request_id')->unsigned();
+                $table->foreign('request_id')
+                        ->references('id')
+                        ->on('requests')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');
+                $table->integer('user_id')->unsigned();
+                $table->foreign('user_id')
+                        ->references('id')
+                        ->on('users')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');
+                $table->timestamps();
+                $table->longtext('details');
+                $table->softDeletes();
+            });
+        } 
     }
 
     /**
