+++ database/migrations/2017_11_13_114934_create_receipts_table.php
@@ -13,28 +13,30 @@ class CreateReceiptsTable extends Migration
      */
     public function up()
     {
-        Schema::create('receipts', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('purchaseorder_id')->unsigned()->nullable();
-            $table->foreign('purchaseorder_id')
-                    ->references('id')
-                    ->on('purchaseorders')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
-            $table->string('number')->unique();
-            $table->string('invoice')->nullable();
-            $table->datetime('invoice_date')->nullable();
-            $table->datetime('date_delivered')->nullable();
-            $table->string('received_by')->nullable();
-            $table->integer('supplier_id')->unsigned()->nullable();
-            $table->foreign('supplier_id')
-                    ->references('id')
-                    ->on('suppliers')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('receipts')){
+            Schema::create('receipts', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('purchaseorder_id')->unsigned()->nullable();
+                $table->foreign('purchaseorder_id')
+                        ->references('id')
+                        ->on('purchaseorders')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');
+                $table->string('number')->unique();
+                $table->string('invoice')->nullable();
+                $table->datetime('invoice_date')->nullable();
+                $table->datetime('date_delivered')->nullable();
+                $table->string('received_by')->nullable();
+                $table->integer('supplier_id')->unsigned()->nullable();
+                $table->foreign('supplier_id')
+                        ->references('id')
+                        ->on('suppliers')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        } 
     }
 
     /**
