+++ database/migrations/2017_11_13_115104_create_receipts_supplies_table.php
@@ -13,26 +13,28 @@ class CreateReceiptsSuppliesTable extends Migration
      */
     public function up()
     {
-        Schema::create('receipts_supplies', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('receipt_id')->unsigned();
-            $table->foreign('receipt_id')
-                    ->references('id')
-                    ->on('receipts')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');        
-            $table->integer('supply_id')->unsigned();
-            $table->foreign('supply_id')
-                    ->references('id')
-                    ->on('supplies')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
-            $table->integer('quantity');
-            $table->integer('remaining_quantity');
-            $table->decimal('unitcost', 8, 2)->nullable();
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('receipts_supplies')){
+            Schema::create('receipts_supplies', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('receipt_id')->unsigned();
+                $table->foreign('receipt_id')
+                        ->references('id')
+                        ->on('receipts')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');        
+                $table->integer('supply_id')->unsigned();
+                $table->foreign('supply_id')
+                        ->references('id')
+                        ->on('supplies')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');
+                $table->integer('quantity');
+                $table->integer('remaining_quantity');
+                $table->decimal('unitcost', 8, 2)->nullable();
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        } 
     }
 
     /**
