+++ database/migrations/2018_01_29_140148_create_rsmi_table.php
@@ -13,26 +13,28 @@ class CreateRsmiTable extends Migration
      */
     public function up()
     {
-        Schema::create('rsmi', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
-            $table->foreign('user_id')
-                    ->references('id')
-                    ->on('users')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
-            $table->string('remarks')->nullable();
-            $table->datetime('report_date')->nullable();
-            //  status can be of the following:
-            //  P - pending
-            //  S - sent by the assets management
-            //  R - received by the accounting
-            //  A - appended to ledger card     
-            //  E - returned by accounting due to problems
-            //  C - cancelled by the assets management 
-            $table->char('status', 2)->nullable();
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('rsmi')){
+            Schema::create('rsmi', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('user_id')->unsigned();
+                $table->foreign('user_id')
+                        ->references('id')
+                        ->on('users')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');
+                $table->string('remarks')->nullable();
+                $table->datetime('report_date')->nullable();
+                //  status can be of the following:
+                //  P - pending
+                //  S - sent by the assets management
+                //  R - received by the accounting
+                //  A - appended to ledger card     
+                //  E - returned by accounting due to problems
+                //  C - cancelled by the assets management 
+                $table->char('status', 2)->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
