+++ database/migrations/2018_03_19_140338_add_user_office_foreign_key.php
@@ -13,13 +13,15 @@ class AddUserOfficeForeignKey extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-                    $table->foreign('office')
-                    ->references('code')
-                    ->on('offices')
-                    ->onUpdate('cascade')
-                    ->onDelete('cascade');
-        });
+        if(!Schema::hasTable('users')){
+            Schema::table('users', function (Blueprint $table) {
+                        $table->foreign('office')
+                        ->references('code')
+                        ->on('offices')
+                        ->onUpdate('cascade')
+                        ->onDelete('cascade');
+            });
+        }
     }
 
     /**
